refractor: moving widget from ui/common to ui/widgets
